chore: apply suggestions Signed-off-by: Sonia Sandler <ssandler@redhat.com>
