--- packages/renderer/src/lib/network/NetworkColumnActions.svelte
@@ -1,12 +1,0 @@
-<script lang="ts">
-import NetworkActions from './NetworkActions.svelte';
-import type { NetworkInfoUI } from './NetworkInfoUI';
-
-interface Props {
-  object: NetworkInfoUI;
-}
-
-let { object }: Props = $props();
-</script>
-
-<NetworkActions network={object} on:update />
