--- packages/renderer/src/lib/network/NetworkColumnDriver.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-import type { NetworkInfoUI } from './NetworkInfoUI';
-
-interface Props {
-  object: NetworkInfoUI;
-}
-
-let { object }: Props = $props();
-</script>
-
-<button
-  class="hover:cursor-pointer flex text-[var(--pd-table-body-text-highlight)] max-w-full overflow-hidden text-ellipsis">
-  {object.driver}
-</button>
