+++ packages/renderer/src/lib/network/columns/NetworkColumnActions.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+import NetworkActions from '../NetworkActions.svelte';
+import type { NetworkInfoUI } from '../NetworkInfoUI';
+
+interface Props {
+  object: NetworkInfoUI;
+}
+
+let { object }: Props = $props();
+</script>
+
+<NetworkActions network={object} />
