+++ packages/renderer/src/lib/network/columns/NetworkColumnId.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import type { NetworkInfoUI } from './NetworkInfoUI';
+import type { NetworkInfoUI } from '../NetworkInfoUI';
 
 interface Props {
   object: NetworkInfoUI;
@@ -10,6 +10,6 @@ let { object }: Props = $props();
 
 <button
   class="hover:cursor-pointer flex text-[var(--pd-table-body-text-highlight)] max-w-full overflow-hidden text-ellipsis"
-  >
-  {object.name}
+ >
+  {object.shortId}
 </button>
